[Feat] 서류 평가, 면접, 최종 선발 플로우 컨트롤러 및 UseCase 제작 (#274)
* feat: 서류 평가 플로우 API 컨트롤러 및 usecase 구현
- 서류 평가 대상 목록 조회 API
- 특정 지원서 상세 조회 API
- 지원서에 대한 운영진 평가 목록 조회
- 지원서에 대한 자신의 평가 조회
- 지원서에 내 평가 등록하기 (POST, PATCH)
- 지원서 합격/불합격 처리
- Evaluation 엔티티에서 EvaluationDecision 삭제

* feat: 면접 스케줄링 플로우 API 컨트롤러 및 usecase 설계
- 면접 스케줄링 summary 조회
- 날짜/파트별 면접 슬롯 조회
- 슬롯 기준 배정 가능한 지원자/배정된 지원자 조회
- 특정 슬롯에 배정된 면접자 목록 조회
- 면접 슬롯 배정/해제

* chore: 서류 평가 관련 항목 파일 이름 및 엔드포인트 수정

* feat: 면접 질문 관련 엔티티 추가

* feat: 면접 질문지 관리 API및 usecase 구현
- 면접 질문 파트 목록 조회
- 파트별 면접 질문 목록 조회
- 면접 질문 등록 / 수정 / 삭제
- 면접 질문 순서 변경

* feat: 면접 질문지 및 면접 평가 컨트롤러/uscase 구현
- 면접용 파트 드롭다운 목록 조회
- 면접 사전 질문 목록 조회
- 면접 평가 대상 목록 조회
- 실시간 면접 평가 목록 조회
- 내 면접 평가 제출 / 재제출
- 내 면접 평가 조회
- 추가 질문 등록 / 수정 / 삭제 / 조회

* rename: 서류 평가 관련 컨트롤러, 서비스 DocEval -> DocumentEvaluation으로 이름 변경

* feat: 서류 평가 POST API를 PATCH API로 통합

* fix: InterviewAssignment에서 status(DRAFT/CONFIRMED) 제거

* fix: 서류 평가 임시저장을 위해 Evaluation에 EvaluationStatus 필드 추가

* fix: 서류 평가 관련 항목 MyEvaluation -> MyDocumentEvaluation으로 변경
- 응답 preferredPart에 priority 추가

* rename: 컨트롤러 이름 수정

* rename: EvaluationStatusCode -> ApplicationEvaluationStatusCode로 이름 변경
- 해당 enum을 지원자의 지원 현황 조회로 한정하기 위함

* feat: 총괄 지원자 관리 리스트 조회 API 컨트롤러/usecaes 구현

* feat: 최종 선발 리스트 조회, 단건 합격/합격취소 API 컨트롤러, usecase 구현

* fix: 서류 평가 대상 목록 조회, 최종 선발 대상 리스트 조회 요청/응답에 pagination 추가
- pagination 관련 변수명 통일

+++ src/main/java/com/umc/product/recruitment/domain/Application.java
@@ -12,6 +12,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.math.BigDecimal;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -44,14 +45,16 @@ public class Application extends BaseEntity {
     @Column(nullable = false)
     private ApplicationStatus status = ApplicationStatus.APPLIED;
 
+    // 운영진의 평가 점수를 평균내어 저장하는 용도
     @Column(name = "doc_score")
-    private Integer docScore;
+    private BigDecimal docScore;
 
     @Column(name = "interview_score")
-    private Integer interviewScore;
+    private BigDecimal interviewScore;
 
+    // 서류 점수 + 면접 점수로 평균내서 저장. 별도의 컬럼으로 저장해놓을 필요가 있을지 검토 필요.
     @Column(name = "final_score")
-    private Integer finalScore;
+    private BigDecimal finalScore;
 
     public static Application createApplied(Recruitment recruitment, Long applicantMemberId, Long formResponseId) {
         return Application.builder()
